@@ -8,7 +8,7 @@
         }
     },
     "require": {
-        "php": ">=5.6.0",
+        "php": "~7.1.0",
         "guzzlehttp/guzzle": "^6.3.0"
     }
 }
