@@ -1,10 +1,10 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace KingApp\SocialQueenApi\Client;
 
 class Channel extends Base
 {
-    public function series()
+    public function list(): array
     {
         return $this->get('channel');
     }
