@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace KingApp\SocialQueenApi\Client;
 
@@ -7,44 +7,44 @@ class Media extends Base
     protected $id;
     protected $path;
 
-    public function series()
+    public function list(): array
     {
         return $this->get('posts');
     }
 
-    public function setId($id)
+    public function setId(string $id): self
     {
         $this->id = $id;
         return $this;
     }
 
-    public function setPath($filePath)
+    public function setPath(string $filePath): self
     {
         $this->path = $filePath;
         return $this;
     }
 
-    public function setRectangle($rectangle)
+    public function setRectangle(array $rectangle): self
     {
         foreach (['x', 'y', 'w', 'h'] as $name) {
             $this->data['rectangle'] = (int)$rectangle[$name];
         }
     }
 
-    public function setLabel($label)
+    public function setLabel(string $label): self
     {
         $this->data['label'] = $label;
         return $this;
     }
 
-    public function toArray()
+    public function toArray(): array
     {
         $params = $this->data;
         $params['uuid'] = $this->id ?: $this->create();
         return $params;
     }
 
-    public function create()
+    public function create(): string
     {
         return $this->id = $this->post('media', ['multipart' => [[
             'name' => 'fileUpload',
