@@ -7,31 +7,31 @@ class Post extends Base
     /** @var Media[] */
     protected $media = [];
 
-    public function setContent($content)
+    public function setContent(string $content): self
     {
         $this->data['content'] = $content;
         return $this;
     }
 
-    public function setTwitterContent($content)
+    public function setTwitterContent(string $content): self
     {
         $this->data['twitter']['content'] = $content;
         return $this;
     }
 
-    public function setTitle($title)
+    public function setTitle(string $title): self
     {
         $this->data['title'] = $title;
         return $this;
     }
 
-    public function setExceedTwitterContent($flag)
+    public function setExceedTwitterContent(bool $flag): self
     {
         $this->data['twitter']['exceed'] = $flag;
         return $this;
     }
 
-    public function setInstagramCredentials($credentials)
+    public function setInstagramCredentials(array $credentials): self
     {
         foreach ($credentials as $uuid => $loginAndPassword) {
             $this->data['channel']['login'][$uuid] = [$loginAndPassword['login'], $loginAndPassword['password']];
@@ -39,7 +39,7 @@ class Post extends Base
         return $this;
     }
 
-    public function setChannels($channels)
+    public function setChannels(array $channels): self
     {
         $this->data['channel']['my'] = [];
         foreach ($channels as $channel) {
@@ -48,7 +48,7 @@ class Post extends Base
         return $this;
     }
 
-    public function setSharedChannels($channels)
+    public function setSharedChannels(array $channels): self
     {
         $this->data['channel']['shared'] = [];
         foreach ($channels as $channel) {
@@ -57,7 +57,7 @@ class Post extends Base
         return $this;
     }
 
-    public function setHashTags($hashTags)
+    public function setHashTags(array $hashTags): self
     {
         $this->data['hashTags'] = [];
         foreach ($hashTags as $hashTag) {
@@ -66,19 +66,19 @@ class Post extends Base
         return $this;
     }
 
-    public function setFacebookLocationId($id)
+    public function setFacebookLocationId(string $id): self
     {
         $this->data['location']['facebookId'] = $id;
         return $this;
     }
 
-    public function setPublishedAt(\DateTime $dateTime)
+    public function setPublishedAt(\DateTime $dateTime): self
     {
         $this->data['publishedAt'] = $dateTime->format(DATE_ATOM);
         return $this;
     }
 
-    public function addMediaFile($fileName, $settings = [])
+    public function addMediaFile(string $fileName, ?array $settings = []): self
     {
         $media = (new Media($this->client));
         $media->setPath($fileName);
@@ -90,7 +90,7 @@ class Post extends Base
         return $this;
     }
 
-    public function setMediaSetting($transformation, $ratio = 0)
+    public function setMediaSetting(string $transformation, ?float $ratio = 0): self
     {
         if (!in_array($transformation, ['c', 'o', 'f'])) throw new \Exception('No such transformation possible use c -auto cropp, f - fill, o -original');
         $this->data['gallery']['setting']['type'] = $transformation;
@@ -100,21 +100,21 @@ class Post extends Base
         return $this;
     }
 
-    public function send()
+    public function send(): string
     {
         if ($this->media && empty($this->data['gallery']['setting'])) throw new \Exception('There is no media settings');
         foreach ($this->media as $media) {
             $this->data['gallery'][spl_object_hash($media)] = $media->toArray();
         }
-        return $this->post('posts', ['form_params' => $this->data])->message;
+        return $this->post('posts', ['form_params' =>  $this->data])->message;
     }
 
-    public function create()
+    public function create(): self
     {
         return new self($this->client);
     }
 
-    public function series()
+    public function list(): array
     {
         return $this->get('posts');
     }
